requirements.txt
opencv-python~=4.5.1.48
matplotlib~=3.4.0
numpy~=1.19.5
scikit-image~=0.18.1
imutils~=0.5.4
tensorflow~=2.4.1
dlib~=19.22.0
segmentation-models~=1.0.1
requests~=2.25.1
PyYAML~=5.4.1
python-telegram-bot~=13.5